third_party/blink/web_tests/fast/js/js-constructors-use-correct-global-expected.txt

Test to ensure that js constructors create objects with the correct constructor

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new otherGlobal.Object() instanceof otherGlobal.Object is true
PASS otherGlobal.Object() instanceof otherGlobal.Object is true
PASS new otherGlobal.Function() instanceof otherGlobal.Function is true
PASS otherGlobal.Function() instanceof otherGlobal.Function is true
PASS new otherGlobal.Array() instanceof otherGlobal.Array is true
PASS otherGlobal.Array() instanceof otherGlobal.Array is true
PASS new otherGlobal.Number() instanceof otherGlobal.Number is true
PASS new otherGlobal.String() instanceof otherGlobal.String is true
PASS new otherGlobal.Boolean() instanceof otherGlobal.Boolean is true
PASS new otherGlobal.RegExp() instanceof otherGlobal.RegExp is true
PASS otherGlobal.RegExp() instanceof otherGlobal.RegExp is true
PASS new otherGlobal.Date() instanceof otherGlobal.Date is true
PASS new otherGlobal.Error() instanceof otherGlobal.Error is true
PASS otherGlobal.Error() instanceof otherGlobal.Error is true
PASS new otherGlobal.RangeError() instanceof otherGlobal.RangeError is true
PASS otherGlobal.RangeError() instanceof otherGlobal.RangeError is true
PASS new otherGlobal.ReferenceError() instanceof otherGlobal.ReferenceError is true
PASS otherGlobal.ReferenceError() instanceof otherGlobal.ReferenceError is true
PASS new otherGlobal.SyntaxError() instanceof otherGlobal.SyntaxError is true
PASS otherGlobal.SyntaxError() instanceof otherGlobal.SyntaxError is true
PASS new otherGlobal.TypeError() instanceof otherGlobal.TypeError is true
PASS otherGlobal.TypeError() instanceof otherGlobal.TypeError is true
PASS new otherGlobal.URIError() instanceof otherGlobal.URIError is true
PASS otherGlobal.URIError() instanceof otherGlobal.URIError is true
PASS new otherGlobal.Image() instanceof otherGlobal.Image is true
PASS successfullyParsed is true

TEST COMPLETE
